frenetix_rl_requirements.txt
frenetix-motion-planner==2024.1
commonroad-route-planner==2022.3
commonroad-drivability-checker==2023.1
frenetix==0.1.3
numpy==1.24.2
omegaconf==2.1.1
scipy==1.9.3
onnxruntime==1.16.3
shapely==2.0.2
wheel==0.38.4
pandas==2.0.3
gymnasium==0.29.1
stable-baselines3[extra]==2.2.1
sb3_contrib==2.2.1
tensorboard==2.15.1
commonroad-io==2023.2
commonroad-vehicle-models==3.0.2